move expected event
